requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
PyJWT>=2.8.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
flask-cors>=3.0.10
passlib[bcrypt]>=1.7.4
sqlalchemy==2.0.46
httpx>=0.26.0
aiosqlite>=0.18.0
pydantic[email]>=2.5.0